requirements.txt
streamlit
yt-dlp
requests
Pillow
